LPS-62327 Update usages
